fix: validate numeric CLI flags in run-eval, reject empty inline flag values
