+++ apps/framework/lib/cli-args.test.ts
@@ -1,5 +1,5 @@
 import { describe, expect, it } from 'vitest';
-import { readFlag } from './cli-args.js';
+import { positiveInteger, readFlag } from './cli-args.js';
 
 describe('readFlag', () => {
   it('reads flags in both --name value and --name=value form', () => {
@@ -11,6 +11,27 @@ describe('readFlag', () => {
     );
     expect(() => readFlag(['--runs', '--other'], 'runs')).toThrow(
       '--runs requires a value'
+    );
+    expect(() => readFlag(['--runs='], 'runs')).toThrow(
+      '--runs requires a value'
+    );
+  });
+});
+
+describe('positiveInteger', () => {
+  it('rejects non-positive-integer CLI options', () => {
+    expect(positiveInteger('3', 'runs')).toBe(3);
+    expect(() => positiveInteger('0', 'runs')).toThrow(
+      '--runs must be a positive integer'
+    );
+    expect(() => positiveInteger('-1', 'runs')).toThrow(
+      '--runs must be a positive integer'
+    );
+    expect(() => positiveInteger('1.5', 'runs')).toThrow(
+      '--runs must be a positive integer'
+    );
+    expect(() => positiveInteger('abc', 'runs')).toThrow(
+      '--runs must be a positive integer'
     );
   });
 });
